chore: refactor project and requires
