+++ application/utils/applicationExceptions.js
